requirements.txt
mediapipe>=0.10.0
opencv-python>=4.8.0
numpy>=1.21.0
mss>=9.0.0
zeroconf>=0.115.0
Pillow>=10.0.0
